Refactor router structure under src
