+++ src/handlers/experience.go
@@ -4,9 +4,9 @@ import (
 	"net/http"
 	"strconv"
 
-	"anonchihaya.co.uk/models"
-	"anonchihaya.co.uk/repositories"
-	"anonchihaya.co.uk/util"
+	"anonchihaya.co.uk/src/models"
+	"anonchihaya.co.uk/src/repositories"
+	"anonchihaya.co.uk/src/util"
 	"github.com/gin-gonic/gin"
 )
