+++ src/routes/post_routes.go
@@ -1,0 +1,33 @@
+package routes
+
+import (
+	"anonchihaya.co.uk/src/handlers"
+	"anonchihaya.co.uk/src/middlewares"
+	"anonchihaya.co.uk/src/repositories"
+	"github.com/gin-gonic/gin"
+)
+
+func registerPostRoutes(r *gin.Engine, key string, postsRepo repositories.PostRepository) {
+	post := r.Group(prefix + "/post")
+	post.Use(middlewares.KeyChecker(key))
+	{
+		post.GET("/latest", func(ctx *gin.Context) {
+			handlers.GetPostLatest(ctx, postsRepo)
+		})
+		post.GET("/project/:id", func(ctx *gin.Context) {
+			handlers.GetPostsShort(ctx, postsRepo)
+		})
+		post.GET("/:id", func(ctx *gin.Context) {
+			handlers.GetPost(ctx, postsRepo)
+		})
+		post.POST("", func(ctx *gin.Context) {
+			handlers.PostPost(ctx, postsRepo)
+		})
+		post.PUT("", func(ctx *gin.Context) {
+			handlers.PutPost(ctx, postsRepo)
+		})
+		post.DELETE("/:id", func(ctx *gin.Context) {
+			handlers.DeletePost(ctx, postsRepo)
+		})
+	}
+}
